fix(reddog): centralize artifact request derivation
Worker-Lane: REDDOG_PROGRESSIVE_POLICY_CHAIN_REGRESSION_PHASE1

Slice: REDDOG_PROGRESSIVE_POLICY_CHAIN_REGRESSION_PHASE1
